commit blindar con Gates

+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Permission\Models\Role;
 use App\Permission\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -15,6 +16,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('haveaccess', 'role.index');
         $roles =  Role::orderBy('id','Desc')->paginate(2);
 
         return view('role.index',compact('roles'));
@@ -27,6 +29,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('haveaccess', 'role.create');
         $permissions = Permission::get();
         return view('role.create', compact('permissions'));
     }
@@ -39,17 +42,18 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('haveaccess', 'role.create');
         $request -> validate([
              'name'        => 'required|max:50|unique:roles,name',
              'slug'        => 'required|max:50|unique:roles,slug',
              'full-access' => 'required|in:yes,no'
         ]);
         $role = Role::create($request->all());
-        if ($request->get('permission')) {
+        //if ($request->get('permission')) {
             //return $request->all();
             $role->permissions()->sync($request->get('permission'));
            
-        }
+        //}
         return redirect()->route('role.index')
             ->with('status_success','Role saved successfully');         
     }
@@ -63,6 +67,7 @@ class RoleController extends Controller
     public function show(Role $role)
     {
         //rol ver
+        $this->authorize('haveaccess', 'role.show');
         $permission_role=[];
 
         foreach($role->permissions as $permission) {
@@ -81,6 +86,7 @@ class RoleController extends Controller
     public function edit(Role $role)
     {
         //Rol de editar
+        $this->authorize('haveaccess', 'role.edit');
         $permission_role=[];
 
         foreach($role->permissions as $permission) {
@@ -100,17 +106,18 @@ class RoleController extends Controller
     public function update(Request $request, Role $role)
     {
         //Rol de actualizar
+        $this->authorize('haveaccess', 'role.edit');
         $request -> validate([
             'name'        => 'required|max:50|unique:roles,name, '.$role->id,            
             'slug'          => 'required|max:50|unique:roles,slug,'.$role->id,
             'full-access' => 'required|in:yes,no'
        ]);
        $role->update($request->all());
-       if ($request->get('permission')) {
+       //if ($request->get('permission')) {
            //return $request->all();
            $role->permissions()->sync($request->get('permission'));
           
-       }
+       //}
        return redirect()->route('role.index')
            ->with('status_success','Role update successfully');
     }
@@ -124,6 +131,7 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         //rol eliminar
+        $this->authorize('haveaccess', 'role.destroy');
         $role->delete();
         return redirect()->route('role.index')
            ->with('status_success','Role successfully removed');
